7/6/2021
Thought about what i am going to do with RL and The current plan is to have random actions with percentage chance of picking an option to take or give a block. Then modifying the weights each game if we find better paths.
I'm going to loop this tomorrow and see what baseline I can get.