7/7/2021
A straight random weighted decision wont work but as expected. I just wanted it to make random choices. 

I think the agents should know what they have and need.   actions that improve the agents status will be given a weight of 1 per say, but if an action is available that leads to asking the other agent to do something then that will be given a weight of 2.