7/7/2021-2
Today I finalized the reward system.
The cliffnotes are basically, there is a weighted random choice for moving , taking  a block or doing nothing.

Depending on if the agent needs a block, owns too many or needs to take an orphan blocks penalties will be given. This took longer than expected but was good to plan out. Tomorrow i will have the first results of the implementation of this logic